==> 4.FST_contemporary.Rmd <==
---
title: "4.FST_intergenic"
output: html_document
---


# Population Genetic Differenciation (Fst)

Our program can estimate fst between populations. 
And has been generalized to give all pairwise fst estimates if you supply the command with multiple populations.
if you supply 3 populations, the program will also output the pbs statistic.

NB we have removed the very unusefull unweighted fst estimator in the output, and have included a header. The output example below will be updated at some point.

The procedure is

1 - Use angsd for calculating saf files for each population
2 - Use realSFS to calculate 2d sfs for each pair
3 - Use the above calculated 2dsfs as priors jointly with all safs from step1 to calculate fst binary files
4 - Use realSFS to extract the the fst values from the fst


## Estimate multidimensional SFS: Calculate all pairwise 2dsfs's

An important issue when doing this is to be sure that we are comparing the exactly same corresponding sites between populations. 
ANGSD does that automatically and considers only a set of overlapping sites. 

First calculate per pop saf for each population. This is already done.


```{r, engine=bash, eval=FALSE}

#SCRIPT TO GENERATE ALL POSSIBLE ITERATIONS
nano pops_combinations.R
========================================================================================================================
#!/usr/bin/env Rscript
library(dplyr)
library(magrittr)
f <- file("stdin")
open(f)
pops <- scan(f, what = "list")
combinations <- as.data.frame(t(combn(pops, 2))) %>% mutate (combination=paste(V1,V2,sep="/")) %>% select(combination)
for (combination in combinations) {
cat(combination)
}
========================================================================================================================
chmod 555 pops_combinations.R


screen -S creating_FST
script pairwise_real2DSFSs_and_Fst.log

RUTA=/home/mlucena/ANGSD_analysis 
ANGSD="/opt/angsd/angsd"
NGSTOOLS="/opt/angsd/angsd/misc"
THREADS=15
SFS_FOLDER=$RUTA/whole_genome_analysis/sfs
FST_FOLDER=$RUTA/whole_genome_analysis/fst

#Get all populations name
cd $SFS_FOLDER
POPS=($(ls *.unfolded-lr.postprob.thetas.gz | sed -e 's/.unfolded-lr.postprob.thetas.gz//g'  | sort | uniq))
cd $FST_FOLDER
#Get all possible pairwise combinations
POP_PAIRS=($(echo ${POPS[@]} | Rscript pops_combinations.R))

# I have 10 combinations:
c_ll_ki_n013/c_ll_no_n008 
c_ll_ki_n013/c_ll_po_n008 
c_ll_ki_n013/c_lp_do_n012 
c_ll_ki_n013/c_lp_sm_n019 
c_ll_no_n008/c_ll_po_n008 
c_ll_no_n008/c_lp_do_n012 
c_ll_no_n008/c_lp_sm_n019 
c_ll_po_n008/c_lp_do_n012 
c_ll_po_n008/c_lp_sm_n019 
c_lp_do_n012/c_lp_sm_n019


#Iterate over combinations
for POP_PAIR in  "${POP_PAIRS[@]}"
do
POP1=($(echo $POP_PAIR | cut -d "/"  -f 1))
POP2=($(echo $POP_PAIR | cut -d "/"  -f 2)) 
echo ">$POP_PAIR"
#Skip iteration if file exists
if [ -f  $POP1.$POP2.unfolded-lr.saf.idx.ml  ]
  then
    echo "PAIR: $POP_PAIR exists."
    continue
fi
echo "--->realSFS 2 dimensional $POP1 and $POP2"
$NGSTOOLS/realSFS -P $THREADS $SFS_FOLDER/$POP1.unfolded-lr.saf.idx $SFS_FOLDER/$POP2.unfolded-lr.saf.idx  > $POP1.$POP2.unfolded-lr.saf.idx.ml

# The output file is a flatten matrix, where each value is the count of sites with the corresponding joint frequency ordered as [0,0] [0,1] and so on.
# less -S KIR.NOR.unfolded.ml.sfs
# No need to plot it, as it a ML estimation. 

# Prepare the fst for easy analysis
echo "-------------------->FST  $POP1 and $POP2"
$NGSTOOLS/realSFS fst index -P $THREADS $SFS_FOLDER/$POP1.unfolded-lr.saf.idx $SFS_FOLDER/$POP2.unfolded-lr.saf.idx  -sfs $POP1.$POP2.unfolded-lr.saf.idx.ml -fstout $POP1.$POP2.unfolded-lr
echo "--------------------------------->FST  $POP1 and $POP2 printing"
#$NGSTOOLS/realSFS fst print $POP1.$POP2.unfolded-lr.fst.idx > $POP1.$POP2.unfolded-lr.fst.print
# Ojo!!! el archivo print ocupa muchísimo!!!!!!
done
```

Esto lo he hecho para todas las poblaciones, pero dos archivos se me han escupido en blanco:
po & sm y do & smo.
He probado a lanzarlos a mano, y me ha dado fallo justo cuando estaba a punto de acabar. Voy a esperar a que el servidor esté un poco más vacío y probar a lanzarlos. 
Lanzados y aparentemente bien. 

## Get the global estimate

```{r, engine=bash, eval=FALSE}

echo "POP1,POP2,FSTunweight,FSTweight" > global.fsts.csv
for POP_PAIR in  ${POP_PAIRS[@]}
do
POP1=($(echo $POP_PAIR | cut -d "/"  -f 1))
POP2=($(echo $POP_PAIR | cut -d "/"  -f 2)) 
FST=($( $NGSTOOLS/realSFS fst stats $POP1.$POP2.unfolded-lr.fst.idx | tr ' ' ',' ))
echo "$POP1,$POP2,$FST" >> global.fsts.csv
done


# you can have a look at their values:
#$NGSTOOLS/realSFS fst print KIR.NOR.BIA.unfolded.pbs.fst.idx | less -S

# where columns are: chromosome, postisition, (a), (a+b) values for the three FST comparisons, where FST is defined as a/(a+b).
# Note that FST on multiple SNPs is calculated as sum(a)/sum(a+b).


# Now, you could do an sliding-window analysis.
# * If you give only 2 pop in input, only the FST will be calculated. 
#### All this analysis could be estimated using NGStools (ngs2dSFS plus ngsFST) but it would need to:

# 1. Define the subset of overlapping sites as it does not compute it automatically as it does ANGSD.
# 2. It could required more memory than ANGSD and may be slower. 
# 3. The methods behind both programs are very similar. 
# 4. ngsTools implements a very simple estimator of the SFS, where the most likely joint allele frequency is recorded at each site. 
# 	 This tends to overestimate singletons, for instance. If you have enough sites, then it is recommended to use the 2DSFS estimated in ANGSD instead. 
# 5. ANGSD can compute a ML estimate of the 2D-SFS which should be preferred when many sites are available. 
# 	 However, ANGSD output file should be transformed (from log to un-log and from space-separated to tab-separated) before being used in ngsFST.


```



###  Copying to local global estimate

```{bash}
scp mlucena@genomics-b.ebd.csic.es:/home/mlucena/ANGSD_analysis/whole_genome_analysis/fst/*.fsts.csv /Users/marialucenaperez/Dropbox/PhD/contemporary/ANGSD/whole_genome_analysis/fst/
```

###  R Neighbour joining tree


```{r}
library(dplyr)
library(tidyr)
library(magrittr)
library(ggplot2)
library(ape)

wd<-"/Users/marialucenaperez/Dropbox/PhD/contemporary/ANGSD/whole_genome_analysis/fst/"

#Look for all global fts table 

fst_table = list.files(path = wd, pattern="*fsts.csv")    


df="global.fsts.csv"

# Define function to plot trees

plot_FST_nj_weighted_unrooted_tree <- function(wd,df,filter_pattern,subset_name )
  {
first.pop <- read.csv( paste0(wd,df) , stringsAsFact=F, sep=",", header = TRUE) %>%
    mutate(pop1.n=(nchar(POP1)-4)/8,pop2.n=(nchar(POP2)-4)/8)%>%
    filter(pop1.n<2, pop2.n<2, grepl(filter_pattern,POP1),grepl(filter_pattern,POP2)) %>% 
    select(POP1) %>% slice(c(1))%>% as.character()
  last.pop<- read.csv( paste0(wd,df) , stringsAsFact=F, sep=",", header = TRUE) %>%
    mutate(pop1.n=(nchar(POP1)-4)/8,pop2.n=(nchar(POP2)-4)/8)%>%
    filter(pop1.n<2, pop2.n<2, grepl(filter_pattern,POP1),grepl(filter_pattern,POP2)) %>% 
    select(POP2) %>% slice(n())%>% as.character()
 
  fst.w_matrix<- read.csv( paste0(wd,df) , stringsAsFact=F, sep=",", header = TRUE) %>%
    rbind(c(first.pop,first.pop,NA,NA),
          c(last.pop,last.pop,NA,NA)) %>%
    mutate(pop1.n=(nchar(POP1)-4)/8,pop2.n=(nchar(POP2)-4)/8)%>%
    filter(pop1.n<2, pop2.n<2, grepl(filter_pattern,POP1),grepl(filter_pattern,POP2)) %>% 
    select(POP1,POP2,FSTweight)%>% 
    spread(POP2,FSTweight)%>%
    tibble::remove_rownames() %>% tibble::column_to_rownames("POP1")
  
  ## plot Neighbouring-Joining Tree
 
    pdf(paste0(wd,subset_name, gsub(".fsts.csv", "_weighted.pdf",df)), paper =  "a4" )
    plot(nj(as.dist(t(fst.w_matrix))), main=gsub(".csv", "_Fst_w", df),sub="neighbor joining",type="unrooted")
    dev.off()
}


# Format tables intro matrices and Plot NJtrees

# Considering the function you can modify the pattern to filter and the name that you gave it. 

# function(wd,df,filter_pattern,subset_name )

# El nombre es "subset_name" que yo escoja, más global.weighted.pdf
  

for (i in 1:length (fst_table)){
  #Load table,  keep only the values of single populations 
  #(I migth be interested in the other comparissons but later)
  plot_FST_nj_weighted_unrooted_tree(wd,fst_table[i],"c_","") 
}



```

## Window analysis


```{bash}

RUTA=/home/mlucena/ANGSD_analysis 
ANGSD="/opt/angsd/angsd"
NGSTOOLS="/opt/angsd/angsd/misc"
THREADS=15
SFS_FOLDER=$RUTA/whole_genome_analysis/sfs
FST_FOLDER=$RUTA/whole_genome_analysis/fst

#Get all populations name
cd $SFS_FOLDER
POPS=($(ls *.unfolded-lr.postprob.thetas.gz | sed -e 's/.unfolded-lr.postprob.thetas.gz//g'  | sort | uniq))
cd $FST_FOLDER
#Get all possible pairwise combinations
POP_PAIRS=($(echo ${POPS[@]} | Rscript pops_combinations.R))


WINDOWSIZE=50000
WINDOWSTEP=10000 

for POP_PAIR in  ${POP_PAIRS[@]}
do
POP1=($(echo $POP_PAIR | cut -d "/"  -f 1))
POP2=($(echo $POP_PAIR | cut -d "/"  -f 2)) 
echo $POP1" & "$POP2" ------> Window analysis"
$NGSTOOLS/realSFS fst stats2 $POP1.$POP2.unfolded-lr.fst.idx -win $WINDOWSIZE -step $WINDOWSTEP  > $POP1.$POP2.unfolded-lr.fst_$WINDOWSIZE.$WINDOWSTEP
done


# Me he dado cuenta que las ventanas no tienen FST en el título, por tanto, tengo que añadirlo. 

FILES_FST=($(ls *10000 | sort | uniq))


for i in "${FILES_FST[@]}"
do
echo $i
cat <(awk -v OFS='\t' '(NR==1) {print $0,"fst"}' $i) <( awk -v OFS='\t' 'NR >= 2 {print $0}' $i) > ${i/.10000/_1000.tsv}
done

# Le dejo el nombre nuevo porque no me gustaba el viejo. 
rm *.10000

```

<!-- ### Copying to local window analysis. -->
<!-- ```{bash} -->
<!-- scp mlucena@genomics-b.ebd.csic.es:/home/mlucena/ANGSD_analysis/whole_genome_analysis/fst/fst_window_analysis/*10000 /Users/marialucenaperez/Dropbox/PhD/contemporary/ANGSD/whole_genome_analysis/fst/ -->
<!-- ``` -->

### R window analysis representation

```{r}


library(dplyr)
library(qqman)
library(ggplot2)
library ("GGally")

wd <- "/home/mlucena/ANGSD_analysis/whole_genome_analysis/fst/fst_window_analysis/"

finsFST = list.files(path = wd, pattern="*fst_50000_1000.tsv$")

for (i in 1:length(finsFST ))
{
  datFST <- read.csv (paste(wd,finsFST [i],sep=""), header = T, sep = '\t')  
  name_FST <- unlist(strsplit(finsFST [i], "[.]"))
  datFST$CHR <- as.numeric(gsub('^.{6}', '', datFST$chr))
  datFST$Nsites <- as.numeric(datFST$Nsites)
  datFST$SNPS_window = paste(datFST$chr, datFST$midPos, sep='_')
  datFST[,-c(1,2)][datFST[, -c(1,2)] < 0] <- 0
  # Manhattan plot FST 
  max_FST <- max (datFST$fst)
  min_FST <- min(datFST$fst)
  setEPS()
  postscript(file=paste(wd,name_FST[1],"-",name_FST[2], '_FST_manhattan.eps', sep=''))
  manhattan(datFST, chr = "CHR", bp = "midPos", p = "fst", snp = "SNPS_window", ylim = c(0 , max_FST+10), ylab = "FST", xlab = "Scaffold", logp=F, genomewideline = F, suggestiveline = F, main = paste ("Manhattan plot of FST-scores ",  name_FST[1],"-",  name_FST[2], sep=''))
  dev.off() 
}


```


```{bash}
scp mlucena@genomics-b.ebd.csic.es:/home/mlucena/ANGSD_analysis/whole_genome_analysis/sfs/window_analysis/*eps /Users/marialucenaperez/Desktop

```









# FST+Theta representation

```{r}

for (i in 1:length(finsFST2pop))
{
  datFST2pop <- read.csv (finsFST2pop[i], header = F, sep = '\t',stringsAsFactors = FALSE, dec = ".")
  name_FST2 <- unlist(strsplit(finsFST2pop[i], "[.]"))
  datFST2pop = datFST2pop[-1,]
  colnames(datFST2pop) <- c("region", "chr", "midPos", "Nsites", "FST")
  datFST2pop$CHR <- as.numeric(gsub('^.{6}', '', datFST2pop$chr))
  datFST2pop$Nsites <- as.numeric(datFST2pop$Nsites)
  datFST2pop$midPos <- as.numeric(datFST2pop$midPos)

datFST2pop[,-c(1,2)][datFST2pop[, -c(1,2)] < 0] <- 0


FST_Thetas_POP1 <- full_join(x = datThetas, y =  datFST2pop, by=c("WinCenter"="midPos", "CHR"="CHR")) %>% 
  #  mutate (normalizedPBS2 = scale(FST_Thetas_POP1$PBS2)) %>% 
  mutate (selectedSNPS = ifelse 
          (tW > filter_tW & FST < filter_FST, "YES", "NO")) %>% 
  mutate (SNPS_window = paste(Chr, WinCenter, sep='_'))

selected_SNPS <- FST_Thetas_POP1 %>% 
  filter (., selectedSNPS == "YES")

snpsOfInterest <- FST_Thetas_POP1[FST_Thetas_POP1$selectedSNPS == "YES", "SNPS_window"]


pdf(file = paste("/Users/marialucenaperez/Dropbox/PhD/contemporary/pdf/", name_FST2[1],"-", name_FST2[2], '_thetas_vs_FST.pdf', sep=''))
ggplot (FST_Thetas_POP1, aes(x=tW, y=FST, fill=selectedSNPS, colour=selectedSNPS)) +
  geom_point() +
  xlab("Waterson estimator") + #x title
  ylab("FST") + # y title
  theme_bw() +  #theme selection for background and lines
  # scale_x_continuous(expand = c(0, 0)) + scale_y_continuous(expand = c(0, 0)) +
  scale_color_manual(values=c("black", "red"))
dev.off()
# MANHATTAN_FST

max_FST2 <- max (datFST2pop$FST)
min_FST2 <- min(datFST2pop$FST)
pdf(file = paste("/Users/marialucenaperez/Dropbox/PhD/contemporary/pdf/", name_FST2[1],"-",name_FST2[2],'_FST.pdf', sep=''))
manhattan(FST_Thetas_POP1, chr = "CHR", bp = "WinCenter", p = "FST", snp = "SNPS_window", ylim = c(min_FST2 , max_FST2+1), ylab = "FST", xlab = "Scaffold", logp=F, genomewideline = F,
          highlight = snpsOfInterest, suggestiveline = FALSE, main = paste("Manhattan plot of FST-scores ",  name_FST2[1],"-",  name_FST2[2], sep=''))
dev.off()
}
#col = c("darkolivegreen3", "dimgrey"),

```
