==> 6.Genetic_distance.Rmd <==
---
title: "6. Genetic distance"
output: html_document
---

## Intergenic data.

```{bash}

cd /home/mlucena/ANGSD_analysis/intergenic_analysis/intergenic_genetic_distances

ANGSD=/opt/angsd/angsd
RUTA=/home/mlucena/ANGSD_analysis 
NGSTOOLS=/opt/ngsTools_19092016


POP="c_ll_no-c_ll_ba-h_ll_ba-c_ll_cr-c_ll_po-c_ll_la-c_ll_ki-c_ll_ur-c_ll_tu-c_ll_to-c_ll_ka-c_ll_og-c_ll_ya-c_ll_vl_n080"

# Running GL:

# En teoría tendría que calcular el archivo de GL con -doGlf 2 -doMajorMinor 1 -doMaf 1. Sin embargo, este ya está calculado puesto que lo usé para NGSAdmix. 
# Por tanto, voy a copiarlo directamente a mi carpeta y todo lo que está ahora mismo "comentado" NO lo he hecho. 

<!-- SNP_PVAL="1e-4" -->
<!-- $ANGSD/angsd -GL 1 -out "$POP"_.genolike -nThreads 40 -doGlf 2 -doMajorMinor 1 -SNP_pval $SNP_PVAL -doMaf 1 -bam $RUTA/intergenic_analysis/"$POP"_n*.intergenic.bamlist -->
<!-- SAMPLE="c_ll_no-c_ll_ba-h_ll_ba-c_ll_cr-c_ll_po-c_ll_la-c_ll_ki-c_ll_ur-c_ll_tu-c_ll_to-c_ll_ka-c_ll_og-c_ll_ya-c_ll_vl_n080.genolike.beagle.gz" -->
<!-- # SAMPLE="$POP"_SingletonSNP_pval_Filtered.genolike.beagle.gz -->
<!-- 
<!-- cat $RUTA/intergenic_analysis/"$POP"_n*.intergenic.bamlist | rev | cut -d "/"  -f 1 | cut -d "_" -f 2,3,4,5 | rev | sed 's/ /n/' > LABELS.txt -->


scp /home/mlucena/ANGSD_analysis/intergenic_analysis/intergenic_ADMIXTURE/*genolike.beagle.gz /home/mlucena/ANGSD_analysis/intergenic_analysis/intergenic_genetic_distances

scp /home/mlucena/ANGSD_analysis/intergenic_analysis/intergenic_ADMIXTURE/*genolike.mafs.gz /home/mlucena/ANGSD_analysis/intergenic_analysis/intergenic_genetic_distances

scp /home/mlucena/ANGSD_analysis/intergenic_analysis/intergenic_ADMIXTURE/*.info /home/mlucena/ANGSD_analysis/intergenic_analysis/intergenic_genetic_distances


N_SITES=`zcat $POP.genolike.mafs.gz | wc -l`
# No tiene cabecero así que no hago tail -n+2

awk '{print $2}' $POP.info > labels.txt


echo $N_SITES 
# 2096552; Perfecto! Igual que PCA y NGSAdmix! :)


# Calculo distancia.

$NGSTOOLS/ngsDist/ngsDist --n_threads 10 --seed 12345 --geno $POP.genolike.beagle.gz --probs --n_ind 80 --n_sites $N_SITES --labels labels.txt --call_geno --out $POP.dist


$NGSTOOLS/ngsDist/ngsDist --n_threads 10 --seed 12345 --geno ../$POP.genolike.beagle.gz --probs --n_ind 80 --n_sites $N_SITES --labels ../labels.txt --call_geno --out $POP.dist



# Ninguno de estos porque ya he hecho la selección de sitios que quiero en el maf. 
# --call_thresh $SNP_PVAL 
# --N_thresh 0.3 

# ../ngsDist --n_threads 10 --seed 12345 --verbose 1 --geno testA_32.geno --probs --n_ind $N_IND --n_sites $N_SITES --labels testA.labels --n_boot_rep 5 --boot_block_size 10 --call_geno --N_thresh 0.3 --call_thresh 0.9 --out testA_32B-10CGf.dist


```


--geno FILE: input file with genotypes, genotype likelihoods or genotype posterior probabilities.
--n_ind INT: sample size (number of individuals).
--n_sites INT: number of sites in input file.
--tot_sites INT: total number of sites in dataset.
--labels FILE: labels, one per line, of the input sequences.
--probs: is the input genotype probabilities (likelihoods or posteriors)?
--log_scale: Ii the input in log-scale?.
--call_geno: call genotypes before running analyses.
--N_thresh DOUBLE: minimum threshold to consider site; missing data if otherwise (assumes -call_geno)
--call_thresh DOUBLE: minimum threshold to call genotype; left as is if otherwise (assumes -call_geno)
--pairwise_del: pairwise deletion of missing data.
--avg_nuc_dist: use average number of nucleotide differences as distance (by default, ngsDist uses genotype distances based on allele frequency differences). Only pairs of heterozygous positions are actually affected when using this option, with their distance being 0.5 (instead of 0 by default).
--indep_geno: assume independence between genotypes?
--n_boot_rep INT: number of bootstrap replicates [0].
--boot_block_size INT: block size for bootstrapping [1].
--out FILE: output file name.
--n_threads INT: number of threads to use. [1]
--version: prints program version and exits.
--verbose INT: selects verbosity level. [1]
--seed INT: random number generator seed (only for the bootstrap analysis).


```{bash}
scp mlucena@genomics-b.ebd.csic.es:/home/mlucena/ANGSD_analysis/intergenic_analysis/intergenic_genetic_distances/*dist /Users/marialucenaperez/Dropbox/PhD/contemporary/ANGSD/intergenic_analysis/analysis_genetic_distance 

scp mlucena@genomics-b.ebd.csic.es:/home/mlucena/ANGSD_analysis/intergenic_analysis/intergenic_genetic_distances/*png /Users/marialucenaperez/Dropbox/PhD/contemporary/ANGSD/intergenic_analysis/analysis_genetic_distance 
```


```{r}

wd <- "/Users/marialucenaperez/Dropbox/PhD/contemporary/ANGSD/intergenic_analysis/analysis_genetic_distance/"
POP="c_ll_no-c_ll_ba-h_ll_ba-c_ll_cr-c_ll_po-c_ll_la-c_ll_ki-c_ll_ur-c_ll_tu-c_ll_to-c_ll_ka-c_ll_og-c_ll_ya-c_ll_vl_n080"

data <- read.table(paste (wd , POP, ".dist", sep=""), sep ="\t", header=F, dec=".",skip = 2) 

# skip 2 para que ignore la primera linea

# Script from http://sebastianraschka.com/Articles/heatmaps_in_r.html
#########################################################
### A) Installing and loading required packages
#########################################################

if (!require("gplots")) {
   install.packages("gplots", dependencies = TRUE)
   library(gplots)
   }
if (!require("RColorBrewer")) {
   install.packages("RColorBrewer", dependencies = TRUE)
   library(RColorBrewer)
   }

#########################################################
### B) Reading in data and transform it into matrix format
#########################################################

rnames <- data[,1]                            # assign labels in column 1 to "rnames"
mat_data <- data.matrix(data[,2:ncol(data)])  # transform column 2-5 into a matrix
rownames(mat_data) <- rnames                  # assign row names


#########################################################
### C) Customizing and plotting the heat map
#########################################################

# creates a own color palette from red to green
my_palette <- colorRampPalette(c("red", "yellow", "green"))(n = 299)

# (optional) defines the color breaks manually for a "skewed" color transition
col_breaks = c(seq(-1,0,length=100),  # for red
  seq(0.01,0.8,length=100),           # for yellow
  seq(0.81,1,length=100))             # for green

# creates a 5 x 5 inch image
png(paste (wd, POP, ".png", sep=""),    # create PNG for the heat map        
  width = 20,        # 5 x 300 pixels
  height =20,
  units = "cm",
  res = 300,            # 300 pixels per inch
  pointsize = 10)        # smaller font size

heatmap.2(mat_data,
  symm = T,
  trace="none",         # turns off trace lines inside the heat map
  main = "Distance matrix", # heat map title
  density.info="none",  # turns off density plot inside color legend
  col=my_palette,     # use on color palette defined earlier
  Colv="NA", # turn off column clustering
  Rowv="NA", 
  dendrogram="none",labCol=data$V1)

<<<<<<< HEAD
=======

distance_nj_matrix <- nj(as.dist(t(mat_data)))

plot.phylo(distance_nj_matrix, type = "u" ,  rotate.tree = -110, lab4ut="axial", edge.width=1)

# then, you will find the tip labels in the following vector:

distance_nj_matrix$tip.label

distance_nj_matrix$tip.label <- data[,1]
distance_nj_matrix$tip.label<- as.character(distance_nj_matrix$tip.label)

plot.phylo(distance_nj_matrix, type = "u" ,  rotate.tree = -110, lab4ut="axial", edge.width=1)

dev.print(dev = pdf, "/Users/marialucenaperez/Dropbox/PhD/contemporary/ANGSD/intergenic_analysis/analysis_genetic_distance/phylo_genetic_distances.pdf")

dev.off()
>>>>>>> 2fb5332ce231e761c269773c6739838c689be38e
```

## Intergenic data blast filtered.

```{bash}

mkdir /home/mlucena/ANGSD_analysis/intergenic_analysis/intergenic_genetic_distances_capture_blasted

cd /home/mlucena/ANGSD_analysis/intergenic_analysis/intergenic_genetic_distances_capture_blasted

POP="c_ll_no-c_ll_ba-h_ll_ba-c_ll_cr-c_ll_po-c_ll_la-c_ll_ki-c_ll_ur-c_ll_tu-c_ll_to-c_ll_ka-c_ll_og-c_ll_ya-c_ll_vl_n080"  # <--CHANGE POP HERE


RUTA=/home/mlucena/ANGSD_analysis/depth_calculus/depth_calculus_intergenic_captured_blasted_samples
THREADS=15                    # no. of computer cores used by bwa and samtools. 20 = OK, >20 = ask people first!

read POP mean sd mean_truncated sd_truncated maxDepth minDepth maxDepth_truncated minDepth_truncated < $RUTA/"${POP}"_mean_sd_depthGlobal_lynx_per_pop_mean_folds_0.95.csv

ANGSD="/opt/angsd/angsd"
NGSTOOLS=/opt/ngsTools_19092016
REF="/home/GRUPOS/grupolince/reference_genomes/lynx_pardinus_genome/lp23_without_repetitive_transposable_low_complexity.fa"
ANC="/home/GRUPOS/grupolince/reference_genomes/lynx_rufus_genome/c_lr_zz_0001_recal1.fa"
FILTER1=" -uniqueOnly 1 -remove_bads 1 -only_proper_pairs 1 -baq 1 -C 50 "
FILTER2=" -minMapQ 30 -minQ 20 -doCounts 1 "
N_IND=$(echo ${POP: -3} )
MIN_IND=$(expr $N_IND / 2)
SNP_PVAL="1e-4"
REGIONFILE="no_genes_Lypa_10000longest_center_final_slop20_dot.rf"

RUTA=/home/mlucena/ANGSD_analysis 

# Sanity checks:

echo $POP
echo $N_IND
echo $MIN_IND
echo $maxDepth
echo $minDepth
echo $SNP_PVAL
# Running GL:

# Para los sitios sin filtrar ya tenía este archivo por el admixture, pero para este caso (las muestras blasted no lo tngo así que lo voy a generar de nuevas).

$ANGSD/angsd -nThreads $THREADS -bam $RUTA/intergenic_analysis/"$POP".intergenic_capture_blast_filtered.bamlist -ref $REF  \
-out "$POP".intergenic_capture_blast_filtered.genolike \
$FILTER1 \
$FILTER2 \
-rf $REGIONFILE \
-GL 1  -doGlf 2 \
-doMajorMinor 1 -doMaf 1 -SNP_pval $SNP_PVAL -skipTriallelic 1 \
-minInd $MIN_IND -setMaxDepth $maxDepth -setMinDepth $minDepth 

# ¡Ojo! Tiene cabecero, así que se lo tengo que quitar para contar el numero de posiciones!!

# Si no, nos da error de que no coinciden!
# cannot read GZip GENO file. Check GENO file and number of sites!

N_SITES=`zcat $POP.intergenic_capture_blast_filtered.genolike.mafs.gz  | tail -n+2 | wc -l`
echo $N_SITES 
# 3694 ¿son muy pocos?

cat /home/mlucena/ANGSD_analysis/intergenic_analysis/$POP.intergenic_capture_blast_filtered.bamlist | rev | cut -d "/"  -f 1 | rev | colrm 13 | cut -d "_" -f 1-3 > $POP.column1.borrar
cat /home/mlucena/ANGSD_analysis/intergenic_analysis/$POP.intergenic_capture_blast_filtered.bamlist | rev | cut -d "/"  -f 1 | rev | colrm 13  > $POP.column2.borrar

paste $POP.column1.borrar $POP.column2.borrar > $POP.intergenic_capture_blast_filtered.info

rm *borrar

awk '{print $2}' $POP.intergenic_capture_blast_filtered.info > labels.txt

# Calculo distancia.

$NGSTOOLS/ngsDist/ngsDist --n_threads 10 --seed 12345 --geno $POP.intergenic_capture_blast_filtered.genolike.beagle.gz --probs --n_ind 80 --n_sites $N_SITES --labels labels.txt --call_geno --out $POP.dist

```

Crear nuevos archivos!!!

```{bash}
scp mlucena@genomics-b.ebd.csic.es:/home/mlucena/ANGSD_analysis/intergenic_analysis/intergenic_genetic_distances_capture_blasted/*dist /Users/marialucenaperez/Dropbox/PhD/contemporary/ANGSD/intergenic_analysis/analysis_genetic_distance_blasted_samples 

scp mlucena@genomics-b.ebd.csic.es:/home/mlucena/ANGSD_analysis/intergenic_analysis/intergenic_genetic_distances_capture_blasted/*png /Users/marialucenaperez/Dropbox/PhD/contemporary/ANGSD/intergenic_analysis/analysis_genetic_distance_blasted_samples 
```


```{r}


library(dplyr)
library(tidyr)
library(magrittr)
library(ggplot2)
library(ape)


wd <- "/Users/marialucenaperez/Dropbox/PhD/contemporary/ANGSD/intergenic_analysis//analysis_genetic_distance_blasted_samples/"
POP="c_ll_no-c_ll_ba-h_ll_ba-c_ll_cr-c_ll_po-c_ll_la-c_ll_ki-c_ll_ur-c_ll_tu-c_ll_to-c_ll_ka-c_ll_og-c_ll_ya-c_ll_vl_n080"

data <- read.table(paste (wd , POP, ".dist", sep=""), sep ="\t", header=F, dec=".",skip = 2) 

# skip 2 para que ignore la primera linea

# Script from http://sebastianraschka.com/Articles/heatmaps_in_r.html
#########################################################
### A) Installing and loading required packages
#########################################################

if (!require("gplots")) {
   install.packages("gplots", dependencies = TRUE)
   library(gplots)
   }
if (!require("RColorBrewer")) {
   install.packages("RColorBrewer", dependencies = TRUE)
   library(RColorBrewer)
   }

#########################################################
### B) Reading in data and transform it into matrix format
#########################################################

rnames <- data[,1]                            # assign labels in column 1 to "rnames"
mat_data <- data.matrix(data[,2:ncol(data)])  # transform column 2-5 into a matrix
rownames(mat_data) <- rnames                  # assign row names

#########################################################
### C) Customizing and plotting the heat map
#########################################################

# creates a own color palette from red to green
my_palette <- colorRampPalette(c("red", "yellow", "green"))(n = 299)

# (optional) defines the color breaks manually for a "skewed" color transition
col_breaks = c(seq(-1,0,length=100),  # for red
  seq(0.01,0.8,length=100),           # for yellow
  seq(0.81,1,length=100))             # for green

# creates a 5 x 5 inch image
png(paste (wd, POP, ".png", sep=""),    # create PNG for the heat map        
  width = 20,        # 5 x 300 pixels
  height =20,
  units = "cm",
  res = 300,            # 300 pixels per inch
  pointsize = 10)        # smaller font size

heatmap.2(mat_data,
  symm = T,
  trace="none",         # turns off trace lines inside the heat map
  main = "Distance matrix", # heat map title
  density.info="none",  # turns off density plot inside color legend
  col=my_palette,     # use on color palette defined earlier
  Colv="NA", # turn off column clustering
  Rowv="NA", 
  dendrogram="none",labCol=data$V1)

dev.off()


distance_nj_matrix <- nj(as.dist(t(mat_data)))

plot.phylo(distance_nj_matrix, type = "u" ,  rotate.tree = -110, lab4ut="axial", edge.width=1)

```
