==> 9.Map_projection_PCA.Rmd <==
---
title: "PCA projection"
output: html_document
---


## Scale Bar

Modify from:

http://editerna.free.fr/wp/?p=76
http://egallic.fr/scale-bar-and-north-arrow-on-a-ggplot2-map/


```{r}
#
# Result #
#--------#
# Return a list whose elements are :
# 	- rectangle : a data.frame containing the coordinates to draw the first rectangle ;
# 	- rectangle2 : a data.frame containing the coordinates to draw the second rectangle ; 


# Yo he eliminado el dos y doble texto.




# 	- legend : a data.frame containing the coordinates of the legend texts, and the texts as well.
#
# Arguments : #
#-------------#
# lon, lat : longitude and latitude of the bottom left point of the first rectangle to draw ;
# distanceLon : length of each rectangle ;
# distanceLat : width of each rectangle ;
# distanceLegend : distance between rectangles and legend texts ;
# dist.units : units of distance "km" (kilometers) (default), "nm" (nautical miles), "mi" (statute miles).

createScaleBar <- function(lon,lat,distanceLon,distanceLat,distanceLegend, dist.units = "km"){
	
  # First rectangle
  
	bottomRight <- gcDestination(lon = lon, lat = lat, bearing = 90, dist = distanceLon, dist.units = dist.units, model = "WGS84")
	
	topLeft <- gcDestination(lon = lon, lat = lat, bearing = 0, dist = distanceLat, dist.units = dist.units, model = "WGS84")
	
	rectangle <- cbind(lon=c(lon, lon, bottomRight[1,"long"], bottomRight[1,"long"], lon),
	lat = c(lat, topLeft[1,"lat"], topLeft[1,"lat"],lat, lat))
	rectangle <- data.frame(rectangle, stringsAsFactors = FALSE)
	

	# Now let's deal with the text
	
max_long <-	rectangle$lon[3]
	
	onTop <- gcDestination(lon = (lon+max_long)/2, lat = lat, bearing = 0, dist = distanceLegend, dist.units = dist.units, model = "WGS84")

	legend <- rbind(onTop)
	legend <- data.frame(cbind(legend, text = c(distanceLon)), stringsAsFactors = FALSE, row.names = NULL)
	return(list(rectangle = rectangle, legend = legend))
}

## We also need a function to obtain the coordinates of the North arrow:

#
# Result #
#--------#
# Returns a list containing :
#	- res : coordinates to draw an arrow ;
#	- coordinates of the middle of the arrow (where the "N" will be plotted).
#
# Arguments : #
#-------------#
# scaleBar : result of createScaleBar() ;
# length : desired length of the arrow ;
# distance : distance between legend rectangles and the bottom of the arrow ;
# dist.units : units of distance "km" (kilometers) (default), "nm" (nautical miles), "mi" (statute miles).
createOrientationArrow <- function(scaleBar, length, distance = 1, dist.units = "km"){
	
  lon <- scaleBar$rectangle[1,1]
	lat <- scaleBar$rectangle[1,2]
	
	# Bottom point of the arrow
	begPoint <- gcDestination(lon = lon, lat = lat, bearing = 0, dist = distance, dist.units = dist.units, model = "WGS84")
	lon <- begPoint[1,"long"]
	lat <- begPoint[1,"lat"]
	
	# Let us create the endpoint
	onTop <- gcDestination(lon = lon, lat = lat, bearing = 0, dist = length, dist.units = dist.units, model = "WGS84")
	
	leftArrow <- gcDestination(lon = onTop[1,"long"], lat = onTop[1,"lat"], bearing = 225, dist = length/5, dist.units = dist.units, model = "WGS84")
	
	rightArrow <- gcDestination(lon = onTop[1,"long"], lat = onTop[1,"lat"], bearing = 135, dist = length/5, dist.units = dist.units, model = "WGS84")
	
	res <- rbind(
			cbind(x = lon, y = lat, xend = onTop[1,"long"], yend = onTop[1,"lat"]),
			cbind(x = leftArrow[1,"long"], y = leftArrow[1,"lat"], xend = onTop[1,"long"], yend = onTop[1,"lat"]),
			cbind(x = rightArrow[1,"long"], y = rightArrow[1,"lat"], xend = onTop[1,"long"], yend = onTop[1,"lat"]))
	
	res <- as.data.frame(res, stringsAsFactors = FALSE)
	
	# Coordinates from which "N" will be plotted
	coordsN <- cbind(x = lon, y = (lat + onTop[1,"lat"])/2)
	
	return(list(res = res, coordsN = coordsN))
}

## The last function enables the user to draw the elements:

#
# Result #
#--------#
# This function enables to draw a scale bar on a ggplot object, and optionally an orientation arrow #
# Arguments : #
#-------------#
# lon, lat : longitude and latitude of the bottom left point of the first rectangle to draw ;
# distanceLon : length of each rectangle ;
# distanceLat : width of each rectangle ;
# distanceLegend : distance between rectangles and legend texts ;
# dist.units : units of distance "km" (kilometers) (by default), "nm" (nautical miles), "mi" (statute miles) ;
# rec.fill, rec2.fill : filling colour of the rectangles (default to white, and black, resp.);
# rec.colour, rec2.colour : colour of the rectangles (default to black for both);
# legend.colour : legend colour (default to black);
# legend.size : legend size (default to 3);
# orientation : (boolean) if TRUE (default), adds an orientation arrow to the plot ;
# arrow.length : length of the arrow (default to 500 km) ;
# arrow.distance : distance between the scale bar and the bottom of the arrow (default to 300 km) ;
# arrow.North.size : size of the "N" letter (default to 6).


# First modify the plot of scale for this particular plot.

scaleBar <- function(lon, lat, distanceLon, distanceLat, distanceLegend, dist.unit = "km", rec.fill = "black", rec.colour = "black", legend.colour = "black", legend.size = 4, orientation = TRUE, arrow.length = 500, arrow.distance = 300, arrow.North.size = 6){
	laScaleBar <- createScaleBar(lon = lon, lat = lat, distanceLon = distanceLon, distanceLat = distanceLat, distanceLegend = distanceLegend, dist.unit = dist.unit)
	# First rectangle
	rectangle1 <- geom_polygon(data = laScaleBar$rectangle, aes(x = lon, y = lat), fill = rec.fill, colour = rec.colour)
	
	# Legend
	scaleBarLegend <- annotate("text", label = paste(laScaleBar$legend[,"text"], dist.unit, sep=""), x = laScaleBar$legend[,"long"], y = laScaleBar$legend[,"lat"], size = legend.size, colour = legend.colour)
	
	res <- list(rectangle1, scaleBarLegend)
	
	if(orientation){# Add an arrow pointing North
		coordsArrow <- createOrientationArrow(scaleBar = laScaleBar, length = arrow.length, distance = arrow.distance, dist.unit = dist.unit)
		arrow <- list(geom_segment(data = coordsArrow$res, aes(x = x, y = y, xend = xend, yend = yend)), annotate("text", label = "N", x = coordsArrow$coordsN[1,"x"], y = coordsArrow$coordsN[1,"y"], size = arrow.North.size, colour = "black"))
		res <- c(res, arrow)
	}
	return(res)
}

```




# PCA projection on map


```{r}

# http://www.molecularecologist.com/2015/04/procrustes-analyses-in-r/

library("png")
library("maps")
library("maptools")
library("ggplot2")
library("grid")
library("ggmap") 
library("mapr")
library("raster")
library("sp")
library("MCMCpack")
library("rworldmap")
library("ggmap")
library("optparse")
library("methods")
library("optparse")
library("dplyr")
library("threejs")
library("RColorBrewer")
require("knitr")
library("rgdal")
library("ggspatial")
library("viridis")

# Now we load our PCA analysis:

WORKING_DIR="/Users/marialucenaperez/Dropbox/PhD/contemporary/ANGSD/intergenic_analysis/analysis_PCA/"


###########################################################################################################################
#######################################       CHOOSE ONE                   ################################################
###########################################################################################################################


# pop=c("c_ll_no-c_ll_ba-h_ll_ba-c_ll_cr-c_ll_po-c_ll_la-c_ll_ki-c_ll_ur-c_ll_tu-c_ll_to-c_ll_ka-c_ll_og-c_ll_ya-c_ll_vl_n080")

# pop=c("c_ll_tu-c_ll_to-c_ll_ka-c_ll_og-c_ll_ya-c_ll_vl_n030")

# pop=c("c_ll_no-c_ll_ba-h_ll_ba-c_ll_cr-c_ll_po-c_ll_la-c_ll_ki-c_ll_ur_n050")


###########################################################################################################################
###########################################################################################################################
###########################################################################################################################



# Read input file
covar <- read.table(paste(WORKING_DIR,pop, ".covar", sep=''), stringsAsFact=F);

# Read annot file
annot <- read.table(paste(WORKING_DIR,pop, ".ALL.clst", sep=""), header=TRUE) 

# Parse components to analyze

comp <- c(1,2)
# comp <- c(2,3)
# comp <- c(1,3)


# Eigenvalues
eig <- eigen(covar, symm=TRUE);
eig$val <- eig$val/sum(eig$val);
cat(signif(eig$val, digits=3)*100,"\n");
eig$values


# Plot
PC <- as.data.frame(eig$vectors)
colnames(PC) <- gsub("V", "PC", colnames(PC))
PC$Pop <- factor(annot$CLUSTER)
PC$Ind <- factor(annot$IID)

coordenadas <- read.csv("/Users/marialucenaperez/Dropbox/tablas_lince/lynxlynx/x_ll_xx_coordinates.csv", sep =";", dec=",")


PC <- PC %>%  mutate(., 
         Subspecies = ifelse (Pop == "c_ll_po" | Pop == "c_ll_ki" | Pop == "c_ll_la" | Pop == "c_ll_no"  | Pop == "c_ll_ur", "L.l.lynx",        
                          ifelse(Pop == "c_ll_ba" | Pop == "h_ll_ba", "L.l.balcanicus" ,
                          ifelse(Pop == "c_ll_cr","L.l.carpathicus",
                          ifelse(Pop == "c_ll_to" | Pop == "c_ll_tu" | Pop == "c_ll_ka", "L.l.kozlovi", 
                          ifelse(Pop == "c_ll_og", "L.l.isabellinus", 
                          ifelse(Pop == "c_ll_vl", "L.l.stroganovi", 
                          ifelse(Pop == "c_ll_ya", "L.l.wrangeli",     
                                 NA)))))))) %>% 
  mutate (., Populations =  ifelse (Pop == "c_ll_po", "NE-Poland",
                          ifelse (Pop == "c_ll_ur", "Urals",
                          ifelse (Pop == "c_ll_ki", "Kirov",
                          ifelse (Pop == "c_ll_la", "Latvia",
                          ifelse (Pop == "c_ll_no", "Norway",
                          ifelse (Pop == "c_ll_ba" | Pop == "h_ll_ba", "Balkans" ,
                          ifelse (Pop == "c_ll_cr","Carpathians",
                          ifelse (Pop == "c_ll_to", "Töv",
                          ifelse (Pop == "c_ll_tu", "Tuva",
                          ifelse (Pop == "c_ll_ka", "Khentii", 
                          ifelse (Pop == "c_ll_og", "Ömnögovi", 
                          ifelse (Pop == "c_ll_vl", "Vladivostok", 
                          ifelse (Pop == "c_ll_ya", "Yakutia",
                          ifelse (Pop == "c_lp_sm", "Sierra Morena",
                          ifelse (Pop == "c_lp_do", "Doñana", NA)))))))))))))))) %>% 
     mutate (., seq = ifelse (Ind == "c_ll_vl_0112" | Ind == "c_ll_vl_0112" | Ind == "c_ll_ya_0146" | Ind == "c_ll_cr_0212" | Ind == "c_ll_ki_0090" , "MACROGEN", "CNAG")) 

cols <- c("NE-Poland"=viridis_pal()(5)[3], 
          "Urals"="#54BAB3", 
          "Balkans"=brewer.pal(12,"Paired")[10], 
          "Carpathians"=brewer.pal(12,"Paired")[9], 
          "Kirov"=viridis_pal()(5)[1], 
          "Latvia"=brewer.pal(12,"Paired")[3], 
       "Norway"=viridis_pal()(5)[2], 
       "Töv"=brewer.pal(12,"Paired")[12], 
          "Tuva"=brewer.pal(12,"Paired")[8], 
          "Khentii"=brewer.pal(12,"Paired")[7], 
          "Ömnögovi"=brewer.pal(12,"Paired")[11], 
          "Vladivostok"=brewer.pal(12,"Paired")[5], 
          "Yakutia"=brewer.pal(12,"Paired")[6],
          "Sierra Morena"=brewer.pal(8, "Greys") [5],
          "Doñana"=brewer.pal(8, "Greys") [8])


# Join PCA with coordinates data:

PC_coordinates <- inner_join(PC,coordenadas, by =c("Ind"="id"))


# To spanish:

PC_coordinates <- PC_coordinates %>% 
    mutate (., Población =  ifelse (Populations == "NE-Poland", "NE-Poland", 
                      ifelse (Populations == "Urals", "Urales", 
                      ifelse (Populations == "Balkans", "Balcanes", 
                      ifelse (Populations == "Carpathians", "Carpatos", 
                      ifelse (Populations == "Kirov", "Kirov", 
                      ifelse (Populations == "Latvia", "Letonia", 
                      ifelse (Populations == "Norway", "Noruega", 
                      ifelse (Populations == "Töv", "Töv", 
                      ifelse (Populations == "Tuva", "Tuva", 
                      ifelse (Populations == "Khentii", "Khentii", 
                      ifelse (Populations == "Ömnögovi", "Ömnögovi", 
                      ifelse (Populations == "Vladivostok", "Vladivostok", 
                      ifelse (Populations == "Yakutia", "Yakutia", NA))))))))))))))


# Create coordinates per pop:

coordenadas_per_pop <- coordenadas %>%  dplyr::mutate (., Populations =  ifelse (pop == "c_ll_po", "NE-Poland",
                          ifelse (pop == "c_ll_ur", "Urals",
                          ifelse (pop == "c_ll_ki", "Kirov",
                          ifelse (pop == "c_ll_la", "Latvia",
                          ifelse (pop == "c_ll_no", "Norway",
                          ifelse (pop == "c_ll_ba" | pop == "h_ll_ba", "Balkans" ,
                          ifelse (pop == "c_ll_cr","Carpathians",
                          ifelse (pop == "c_ll_to", "Töv",
                          ifelse (pop == "c_ll_tu", "Tuva",
                          ifelse (pop == "c_ll_ka", "Khentii", 
                          ifelse (pop == "c_ll_og", "Ömnögovi", 
                          ifelse (pop == "c_ll_vl", "Vladivostok", 
                          ifelse (pop == "c_ll_ya", "Yakutia",
                          ifelse (pop == "c_lp_sm", "Sierra Morena",
                          ifelse (pop == "c_lp_do", "Doñana", NA)))))))))))))))) %>%  dplyr::group_by(Populations) %>%  dplyr::summarise(latitude = mean(latitude), longitude=mean(longitude))



coordenada_per_pop <- coordenadas_per_pop %>%  group_by(Populations) %>%
  summarize(mean_latitude = mean(latitude, na.rm = TRUE), mean_longitude = mean(longitude, na.rm = TRUE))


# As matrix for procrustes analysis.

coord_matrix<-as.matrix(cbind(PC_coordinates$longitude,PC_coordinates$latitude))
PC_matrix<-as.matrix(cbind(PC_coordinates$PC1,PC_coordinates$PC2))

# Do procrustes analysis.

p <-procrustes(PC_matrix,coord_matrix,translation=TRUE,dilation=TRUE)

# Cbind with all the infor for the plot.

PCA_procrustes <- data.frame(p$X.new) %>% cbind (., dplyr::select(PC,Ind,Populations))

write.csv(PCA_procrustes, "/Users/marialucenaperez/Owncloud/publico/Eurasian_lynx_phylogeography/procrustes_map.csv", row.names =F) 


# Order factors for legend:

PCA_procrustes$Populations <- factor (PCA_procrustes$Populations, levels=c("Norway", "Balkans", "Carpathians", "NE-Poland", "Latvia", "Kirov", "Urals", "Tuva", "Töv", "Khentii", "Ömnögovi",  "Yakutia", "Vladivostok"))
PC_coordinates$Populations <- factor (PCA_procrustes$Populations, levels=c("Norway", "Balkans", "Carpathians", "NE-Poland", "Latvia", "Kirov", "Urals", "Tuva", "Töv", "Khentii", "Ömnögovi",  "Yakutia", "Vladivostok"))


# Fisrt I get the limits of my PCA and my sampling:

max_long_PCA=max(PCA_procrustes$X1)
min_long_PCA=min(PCA_procrustes$X1)

max_lat_PCA=max(PCA_procrustes$X2)
min_lat_PCA=min(PCA_procrustes$X2)


max_long_sampling=max(PC_coordinates$longitude)
min_long_sampling=min(PC_coordinates$longitude)

max_lat_sampling=max(PC_coordinates$latitude)
min_lat_sampling=min(PC_coordinates$latitude)


# Get KML layer

distribution_layer <- readOGR("/Users/marialucenaperez/Owncloud/publico/Phd/Eurasian_lynx_phylogeography/Lynx_lynx distribution.kml")

# North png file

img <- readPNG("/Users/marialucenaperez/Owncloud/publico/Images_Maps/imagenes/North.png")
north_image <- rasterGrob(img, interpolate=TRUE)



############################################  GENERAL MAP #######################################

max_long=max_long_sampling
min_long=min_long_sampling
max_lat=max_lat_sampling 
min_lat=min_lat_sampling 


general_map <- ggplot() +
  borders(database = "world", fill="gray95", color="gray95", xlim = c(min_long-20, max_long+20),ylim = c(min_lat-20, max_lat+20))+
  #geom_spatial_polygon(data=distribution_layer, aes(fill="gray85", alpha=0.5, inherit.aes = NULL, show.legend = F)) +
  theme (legend.position="none",panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
panel.background = element_rect(fill="white"), axis.title = element_text(size = 20), axis.text = element_text(size=18), title=element_text(size=25)) +
  geom_point(data=coordenadas_per_pop, aes(x=longitude, y=latitude, fill=Populations), size=2, shape=20) + scale_fill_manual(values = cols) +
  
  # Values for whole world 
coord_map (xlim = c(min_long-20, max_long+20),ylim = c(min_lat-13, max_lat+13)) +
scaleBar(lon = 146, lat = 30, distanceLon = 1000, distanceLat = 10, distanceLegend = 230, dist.unit = "km", orientation = FALSE) 

general_map2 <- cowplot::ggdraw(general_map ) + cowplot::draw_grob(north_image ,x=0.95, y=0.28, width=0.05, height=0.05)

ggsave(paste("/Users/marialucenaperez/Owncloud/publico/Eurasian_lynx_phylogeography/maps/",pop,"_general_map.pdf", sep=""), general_map2, device = pdf, width = 15.1, height =  13.7 )

# Whole world : 15.1 x 13.7 


############################################  GENERAL MAP - ONLY ONE COORDINATE PER POP #######################################

max_long=max_long_sampling
min_long=min_long_sampling
max_lat=max_lat_sampling 
min_lat=min_lat_sampling 


general_map <- ggplot() +
  borders(database = "world", fill="gray95", color="gray95", xlim = c(min_long-20, max_long+20),ylim = c(min_lat-20, max_lat+20))+
  geom_spatial(distribution_layer, fill="gray85", alpha=0.5, inherit.aes = NULL, show.legend = F) +
  theme (legend.position="none",panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
panel.background = element_rect(fill="white"), axis.title = element_blank(), axis.text = element_text(size=18), title=element_text(size=25)) +
  geom_point(data=coordenada_per_pop, aes(x=mean_longitude, y=mean_latitude, fill=Populations), size=2, shape=20) + scale_fill_manual(values = cols) +
  
# Values for whole world 
coord_map (xlim = c(min_long-20, max_long+20),ylim = c(min_lat-13, max_lat+13)) +
scaleBar(lon = 146, lat = 30, distanceLon = 1000, distanceLat = 10, distanceLegend = 230, dist.unit = "km", orientation = FALSE) 

general_map2 <- cowplot::ggdraw(general_map ) + cowplot::draw_grob(north_image ,x=0.95, y=0.28, width=0.05, height=0.05)

ggsave(paste("/Users/marialucenaperez/Owncloud/publico/Eurasian_lynx_phylogeography/maps/",pop,"_general_map_only_one_coordinate.pdf", sep=""), general_map2, device = pdf, width = 15.1, height =  13.7 )

# Whole world : 15.1 x 13.7 





########################################################################################################## 
########################################################################################################## 




############################################  SAMPLING MAP #######################################

max_long=max_long_sampling
min_long=min_long_sampling
max_lat=max_lat_sampling 
min_lat=min_lat_sampling 

########################################################################################################## 
########################################################################################################## 

sampling_map <- ggplot() +
  borders(database = "world", fill="gray95", color="gray95", xlim = c(min_long-20, max_long+20),ylim = c(min_lat-20, max_lat+20))+
 # geom_spatial(distribution_layer, fill="gray85", alpha=0.5, inherit.aes = NULL, show.legend = F) +
  theme (legend.justification="top", panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
panel.background = element_rect(fill="white"), legend.text=element_text(size=18),legend.title = element_text(size=20), axis.title = element_text(size = 20), axis.text = element_text(size=18), title=element_text(size=25),  
legend.key = element_rect(colour = 'white', fill = 'white', inherit.blank = T)) +
  geom_point(data=PC_coordinates, aes(x=longitude, y=latitude, fill=Populations), size=5, shape=21, color="dimgrey", pch=15, inherit.aes = F) +
   scale_fill_manual("Población",
  labels = c("Noruega","Balcanes", "Carpatos", "Bialowieza","Letonia","Kirov","Urales","Tuva", "Töv","Khentii","Ömnögovi", "Vladivostok","Yakutia"),
  values = cols) + xlab ("Longitud") + ylab ("Latitud") + 
########################################################################################################## 
#########################################       CHANGE      ##############################################   
########################################################################################################## 
# Whole
# coord_map (xlim = c(min_long-20, max_long+20),ylim = c(min_lat-13, max_lat+13)) +
# scaleBar(lon = 146, lat = 30, distanceLon = 1000, distanceLat = 10, distanceLegend = 230, dist.unit = "km", orientation = FALSE) 
# sampling_map2 <- cowplot::ggdraw(sampling_map) + cowplot::draw_grob(north_image ,x=0.95, y=0.28, width=0.05, height=0.05)

# East
 coord_map (xlim = c(min_long-10, max_long+10),ylim = c(min_lat-10, max_lat+10)) +
scaleBar(lon = 140, lat = 35, distanceLon = 1000, distanceLat = 10, distanceLegend = 200, dist.unit = "km", orientation = FALSE) 
# sampling_map2 <- cowplot::ggdraw(sampling_map) + cowplot::draw_grob(north_image ,x=0.85, y=0.13, width=0.05, height=0.05)

# West
# coord_map (xlim = c(min_long-10, max_long+10),ylim = c(min_lat-10, max_lat+10)) +
# scaleBar(lon = 58, lat = 35, distanceLon = 1000, distanceLat = 10, distanceLegend = 200, dist.unit = "km", orientation = FALSE) 
# sampling_map2 <- cowplot::ggdraw(sampling_map) + cowplot::draw_grob(north_image ,x=0.93, y=0.20, width=0.05, height=0.05)

########################################################################################################## 

ggsave(paste0("/Users/marialucenaperez/Owncloud/publico/PhD/Eurasian_lynx_phylogeography/maps/all_pops_sampling_map.pdf"), plot=sampling_map, width = 30, height = 20)
# Whole world : 15.1 x 13.7 
# West: Saving 12.9 x 13.7 in image

########################################################################################################## 
########################################################################################################## 







############################################  PROCRUSTES MAP #######################################

max_long=max(max_long_PCA,max_long_sampling )
min_long=min(min_long_PCA,min_long_sampling)
max_lat=max(max_lat_PCA,max_lat_sampling )
min_lat=min(min_lat_PCA,min_lat_sampling )


title <- paste("PC",comp[1]," (",signif(eig$val[comp[1]], digits=3)*100,"%)"," / PC",comp[2]," (",signif(eig$val[comp[2]], digits=3)*100,"%)",sep="",collapse="")

PCA_map <- ggplot() +
  borders(database = "world", fill="gray95", color="gray95", xlim = c(min_long-20, max_long+20),ylim = c(min_lat-20, max_lat+20))+
 # geom_spatial(distribution_layer, fill="gray85", alpha=0.5, inherit.aes = NULL, show.legend = F) +
  theme (legend.position="left", panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
  panel.background = element_rect(fill="white"), legend.text=element_text(size=18),legend.title = element_text(size=20), axis.title = element_text(size = 20), axis.text = element_text(size=18),   title=element_text(size=25),  
  legend.key = element_rect(colour = 'white', fill = 'white', inherit.blank = T)) +
  geom_point(data = PCA_procrustes, aes (X1, X2,  fill=Populations), color="dimgrey", size=3, shape=23) +
  ggtitle(title) + 
  coord_map("gilbert", xlim = c(min_long-20, max_long+20),ylim = c(min_lat-20, max_lat+20))+
  scale_fill_manual("Populations",
# labels = c("Noruega","Balcanes", "Carpatos", "Bialowieza","Letonia","Kirov","Urales","Tuva", "Töv","Khentii","Ömnögovi", "Vladivostok","Yakutia"),
  values = cols) +  
########################################################################################################## 
#########################################       CHANGE      ##############################################   
########################################################################################################## 
# Whole
# coord_cartesian (xlim = c(min_long-20, max_long+20),ylim = c(min_lat-20, max_lat+20)) +
# scaleBar(lon = 148, lat = 25, distanceLon = 1000, distanceLat = 10, distanceLegend = 230, dist.unit = "km", orientation = FALSE) 
# PCA_map2 <- cowplot::ggdraw(PCA_map) + cowplot::draw_grob(north_image ,x=0.95, y=0.28, width=0.05, height=0.05)

# East  
#scale_x_continuous(min_long-10, max_long+10)+  scale_y_continuous(min_lat-10, max_lat+10)+
#coord_cartesian (xlim = c(min_long-10, max_long+10),ylim = c(min_lat-10, max_lat+10)) +
scaleBar(lon = 138, lat = 27, distanceLon = 1000, distanceLat = 10, distanceLegend = 200, dist.unit = "km", orientation = FALSE) 
PCA_map2 <- cowplot::ggdraw(PCA_map) + cowplot::draw_grob(north_image ,x=0.89, y=0.13, width=0.05, height=0.05)

# West
#coord_map (xlim = c(min_long-10, max_long+10),ylim = c(min_lat-10, max_lat+10)) +
#scaleBar(lon = 59, lat = 33, distanceLon = 1000, distanceLat = 10, distanceLegend = 100, dist.unit = "km", orientation = FALSE) 
#PCA_map2 <- cowplot::ggdraw(PCA_map) + cowplot::draw_grob(north_image ,x=0.93, y=0.20, width=0.05, height=0.05)

########################################################################################################## 

ggsave(paste("/Users/marialucenaperez/Owncloud/publico/PhD/Eurasian_lynx_phylogeography/maps/",pop,"_PCA_map.pdf", sep=""),  PCA_map)




# -------
# PROCRUSTES & SAMPLING


title <- paste("PC",comp[1]," (",signif(eig$val[comp[1]], digits=3)*100,"%)"," / PC",comp[2]," (",signif(eig$val[comp[2]], digits=3)*100,"%)",sep="",collapse="")


ggplot() +
  borders(database = "world", fill="gray95", color="gray95", xlim = c(min_long-20, max_long+20),ylim = c(min_lat-20, max_lat+20))+
  geom_point(data = PC_coordinates, aes(x=longitude, y=latitude, fill=Populations), size=5, shape=21, color="dimgrey", pch=15, inherit.aes = F) +
  geom_point(data = PCA_procrustes, aes (X1, X2,  fill=Populations), color="dimgrey", size=3, shape=23) +
  theme (legend.position="left", panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
  panel.background = element_rect(fill="white"), legend.text=element_text(size=18),legend.title = element_text(size=20), axis.title = element_text(size = 20), axis.text = element_text(size=18),   title=element_text(size=25),  
  legend.key = element_rect(colour = 'white', fill = 'white', inherit.blank = T)) +
  ggtitle(title) + 
  coord_map("gilbert", xlim = c(min_long-20, max_long+20),ylim = c(min_lat-20, max_lat+20))+
  scale_fill_manual("Populations",
  values = cols) +  
scaleBar(lon = 138, lat = 27, distanceLon = 1000, distanceLat = 10, distanceLegend = 200, dist.unit = "km", orientation = FALSE) +
 # geom_segment(data = merge(PC_coordinates, PCA_procrustes, by = 'Ind'), aes(x = longitude, xend = X1, y = latitude, yend = X2)) +
  guides(fill = FALSE) +
ggsave(paste("/Users/marialucenaperez/Owncloud/publico/PhD/Eurasian_lynx_phylogeography/maps/",pop,"_PCA_map_with_sampling_data_no_lines.pdf", sep=""), device="pdf")

```
```{r}


library(dplyr)
library(ggplot2)
library(sp)
library(maptools)
library(mapproj)
library(magrittr)

# Maptools dataset
data(wrld_simpl)
world <- fortify(wrld_simpl)



worldmerc <-  SpatialPointsDataFrame(coords = data_frame(x = world$long, y = world$lat), 
                                data = world, proj4string = CRS("+proj=longlat +datum=WGS84")) %>%
         subset((lat < 90 & lat > -90)) %>%   # needed because transform not defined at the poles !!!!
         spTransform(CRS("+init=epsg:3857")) 
worldmerc  <-  mutate(worldmerc@data, longmerc = coordinates(worldmerc)[,1], latmerc = coordinates(worldmerc)[,2])


xlim = c(40,150)
ylim = c(30,80)


# Get the coordinates of the limits in mercator projection
lims = SpatialPoints(coords = data_frame(x = xlim, y = ylim), 
                     proj4string = CRS("+proj=longlat +datum=WGS84"))%>%
       spTransform(CRS("+init=epsg:3857"))

# Create regular "grids" of latlon coordinates and find points 
# within xlim/ylim - will be our labels

majgrid_wid_lat = 20
majgrid_wid_lon = 30

majbreaks_lon = data_frame(x=seq(-180,  180, majgrid_wid_lon)) %>% 
                filter(x >= xlim[1] & x <= xlim[2]) %>% 
                as.data.frame()
majbreaks_lat = data_frame(x=seq(-90,   90, majgrid_wid_lat)) %>%
                filter(x >= ylim[1] & x <= ylim[2]) %>% 
                as.data.frame()

#Find corresponding mercator coordinates

mercbreaks_lat = SpatialPoints(coords = expand.grid(x = majbreaks_lon$x, y = majbreaks_lat$x), proj4string = CRS("+init=epsg:4326")) %>% spTransform(CRS("+init=epsg:3857")) %>% coordinates() %>% extract(,2) %>% unique() 

mercbreaks_lon = SpatialPoints(coords = expand.grid(x = majbreaks_lon$x, y = majbreaks_lat$x), proj4string = CRS("+init=epsg:4326"))%>% spTransform(CRS("+init=epsg:3857")) %>% coordinates()  %>% extract(,1) %>% unique()



ggplot(worldmerc, mapping = aes(x = longmerc, y = latmerc, group = group)) +
  geom_polygon(fill = "gray95", colour = "gray95") + 
  coord_fixed(xlim = coordinates(lims)[,1], ylim = coordinates(lims)[,2]) +
  scale_x_continuous("lon", breaks = mercbreaks_lon, labels = signif(majbreaks_lon$x, 2)) + 
  scale_y_continuous("lat", breaks = mercbreaks_lat, labels = signif(majbreaks_lat$x,2)) +
  geom_spatial(distribution_layer, fill="red", alpha=0.5, inherit.aes = NULL, show.legend = F) +
    theme (legend.position="left", panel.grid.major = element_blank(), panel.grid.minor = element_blank(),  
           panel.background = element_rect(fill="white"), legend.text=element_text(size=18),legend.title = element_text(size=20),
           axis.title = element_text(size = 20), axis.text = element_text(size=18),   title=element_text(size=25),  
           legend.key = element_rect(colour = 'white', fill = 'white', inherit.blank = T))






coord_cartesian (xlim = c(min_long-10, max_long+10),ylim = c(min_lat-10, max_lat+10)) +
scaleBar(lon = 138, lat = 27, distanceLon = 1000, distanceLat = 10, distanceLegend = 200, dist.unit = "km", orientation = FALSE) 
PCA_map2 <- cowplot::ggdraw(PCA_map) + cowplot::draw_grob(north_image ,x=0.89, y=0.13, width=0.05, height=0.05)

  borders(database = "world", fill="gray95", color="gray95", xlim = c(min_long-20, max_long+20),ylim = c(min_lat-20, max_lat+20))+
  geom_spatial(distribution_layer, fill="gray85", alpha=0.5, inherit.aes = NULL, show.legend = F) +
  theme (legend.position="left", panel.grid.major = element_blank(), panel.grid.minor = element_blank(),
  panel.background = element_rect(fill="white"), legend.text=element_text(size=18),legend.title = element_text(size=20), axis.title = element_text(size = 20), axis.text = element_text(size=18),   title=element_text(size=25),  
  legend.key = element_rect(colour = 'white', fill = 'white', inherit.blank = T)) +
  geom_point(data = PCA_procrustes, aes (X1, X2,  fill=Populations), color="dimgrey", size=3, shape=23) +
  ggtitle(title) + 


```
