==> 15.GC_content.Rmd <==
---
title: "15.GC_content"
output: html_document
---

# GC content calculation

Calculo la GC content basandome en esta discusión
https://www.biostars.org/p/47047/

## Lista de unidades

Las copio del script de divergencia:

```{bash}

mkdir GC_content
cd /home/mlucena/GC_content/

scp /home/mlucena/divergence/list_of_units_with_chr_info_in_lynx_no_duplicates_per_scaffold/list_of_units_with_chr_info_in_lynx_no_duplicates.bed .

# Como hice un sort en la ultima linea tiene el cabecero:

head -n -1 list_of_units_with_chr_info_in_lynx_no_duplicates.bed > temp.bed
mv temp.bed list_of_units_with_chr_info_in_lynx_no_duplicates.bed

bedtools nuc -fi /GRUPOS/grupolince/reference_genomes/lynx_pardinus_genome/lp23.fa -bed /home/mlucena/GC_content/list_of_units_with_chr_info_in_lynx_no_duplicates.bed > list_of_units_with_chr_info_in_lynx_no_duplicates.GC_content.lp.bed
bedtools nuc -fi /GRUPOS/grupolince/reference_genomes/lynx_lynx_genome/LynxLynx_SNVsOnly.fa -bed /home/mlucena/GC_content/list_of_units_with_chr_info_in_lynx_no_duplicates.bed > list_of_units_with_chr_info_in_lynx_no_duplicates.GC_content.ll.bed

```


# Download GC content file

```{bash}

scp  mlucena@genomics-b.ebd.csic.es:/home/mlucena/GC_content/list_of_units_with_chr_info_in_lynx_no_duplicates.GC_content.lp.bed /Users/marialucenaperez/Owncloud/publico/PhD/WG_diversity/GC_content
scp  mlucena@genomics-b.ebd.csic.es:/home/mlucena/GC_content/list_of_units_with_chr_info_in_lynx_no_duplicates.GC_content.ll.bed /Users/marialucenaperez/Owncloud/publico/PhD/WG_diversity/GC_content


```
